requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
selenium-wire>=5.0.0
webdriver-manager>=3.8.0
beautifulsoup4>=4.11.0
requests>=2.28.0
tqdm>=4.64.0 